use shared wrapper for all math skills pages
