+++ apps/web/src/components/math-skills/number-line-exercise/number-line-exercise.tsx
@@ -70,7 +70,7 @@ export function NumberLineExercise() {
   }, [selectedValue, searchedValue, isChecked])
 
   return (
-    <div className="relative mx-4 my-16 max-w-lg bg-white">
+    <>
       <h2 className="pb-8 text-left text-2xl font-bold">
         Wo ist die <span className="text-newgreen">{searchedValue}</span>?
       </h2>
@@ -137,6 +137,6 @@ export function NumberLineExercise() {
           )}
         </div>
       </div>
-    </div>
+    </>
   )
 }
